Implement Customer Auth (Login and Register)
