+++ src/app/customer/auth/auth.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { AuthController } from './auth.controller';
+
+@Module({
+  controllers: [AuthController],
+})
+export class CustomerAuthModule {}
