+++ src/app/customer/customer.controller.ts
@@ -1,6 +1,5 @@
 import { Me } from '@core/decorators/user.decorator';
 import { AuthGuard } from '@core/guards/auth.guard';
-import { CustomerService } from '@core/services/customer.service';
 import { Customer, CustomerStatus } from '@db/entities/core/customer.entity';
 import { RawTransformer } from '@db/transformers/raw.transformer';
 import { NotPermittedException } from '@lib/exceptions/not-permitted.exception';
@@ -12,39 +11,13 @@ import { BadRequestException, Body, Controller, Get, Post, Res, UseGuards } from
 
 @Controller()
 export class CustomerController {
-  constructor(private custService: CustomerService) {}
-
   @Get('me')
   @UseGuards(AuthGuard())
   async me(@Me() me, @Res() response) {
     return response.item(me, RawTransformer);
   }
 
-  @Post('/register')
-  async register(@Body() body, @Res() response) {
-    const rules = {
-      name: 'required|safe_text',
-      phone: 'required|phone|unique',
-      email: 'email',
-    };
-    const validation = Validator.init(body, rules);
-    if (validation.fails()) {
-      throw new ValidationException(validation);
-    }
-
-    const isExist = await Customer.exists({ where: [{ phone: body.phone }, { email: body.email }] });
-
-    if (isExist) {
-      throw new BadRequestException('Phone number is already exist. Try to sign in');
-    }
-
-    const customer = await this.custService.register(body);
-    const token = await this.custService.login(customer);
-
-    return response.data(token);
-  }
-
-  @Post('/verify')
+  @Post('/me/verify')
   @UseGuards(AuthGuard())
   async verify(@Body() body, @Res() response, @Me() me: Customer) {
     const rules = {
@@ -77,7 +50,7 @@ export class CustomerController {
     }
   }
 
-  @Post('/request-otp')
+  @Post('/me/request-otp')
   @UseGuards(AuthGuard())
   async postRequestOTP(@Res() response, @Me() me: Customer) {
     if (me.status !== CustomerStatus.Verify) {
