+++ src/app/customer/customer.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { CustomerAuthModule } from './auth/auth.module';
 import { CustomerController } from './customer.controller';
 import { CustomerOrderModule } from './order/order.module';
 
 @Module({
-  imports: [CustomerOrderModule],
+  imports: [CustomerAuthModule, CustomerOrderModule],
   controllers: [CustomerController],
   providers: [],
 })
